# api/openapi.yaml
openapi: 3.0.3
info:
  title: Anotify API
  version: "0.1.0"
  description: |
    Anotify 后端 API。静态资源走 /static/*（CDN），动态 API 走 /v1/*（no-store）。
    鉴权两种方式：
      - 会话 Cookie（Passkey 登录后的工作台浏览）
      - Bearer API Key（Agent 上报 / 接收端长连接），Key 仅存哈希，带 scope
servers:
  - url: /v1
...
paths:
...
  /keys:
    get:
      tags: [keys]
      summary: 列出 API Keys（不含明文/哈希）
      responses:
        "200": { description: Key 列表 }
    post:
      tags: [keys]
      summary: 新建 API Key（仅此处返回一次明文）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string }
                scopes:
                  type: array
                  items:
                    {
                      type: string,
                      enum: [notify:send, notify:receive, devices:read],
                    }
      responses:
        "200": { description: 创建的 Key（含明文，仅此一次） }
...
  /devices:
    get:
      tags: [devices]
      summary: 列出接收设备
      responses:
        "200": { description: 设备列表 }
    post:
      tags: [devices]
      summary: 注册/更新推送订阅设备（前端订阅后上报）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceUpsert" }
      responses:
        "200": { description: 已保存 }
...
  /notify:
    post:
      tags: [notify]
      summary: Agent 上报一条通知（Bearer Key，scope=notify:send）
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NotifyRequest" }
      responses:
        "200":
          description: 投递结果（含每台设备状态）
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  matched: { type: integer }
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        device: { type: string }
                        status: { type: string }
                        error: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /test-notify:
    post:
      tags: [notify]
      summary: 工作台发送测试通知（会话 Cookie 鉴权）
      description: |
        已登录用户在网页工作台直接发一条测试通知到自己所有已启用且匹配的设备，
        无需持有 send Key。userID 从会话取，消息以该用户身份发布。
        payload 带 source="test-notify" 标记，便于在通知历史里区分测试消息。
      security: [{ cookieAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TestNotifyRequest" }
      responses:
        "200":
          description: 投递结果（含每台设备状态）
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  matched: { type: integer }
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        device: { type: string }
                        status: { type: string }
                        error: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /notifications:
    get:
      tags: [notifications]
      summary: 通知历史（分页）
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - {
            name: before,
            in: query,
            schema: { type: string },
            description: 游标(message id),
          }
      responses:
        "200": { description: 历史列表 }
...
  /stream:
    get:
      tags: [stream]
      summary: WebSocket 长连接接收通知（Bearer Key，scope=notify:receive）
      description: |
        升级为 WebSocket。帧协议（JSON）：
          服务端→客户端: hello / notification / replay_end / pong / subscribed / error / bye
          客户端→服务端: subscribe / unsubscribe / ack / ping / resume
        断线重连用 Last-Event-Id 头或 resume 帧携带 resume_token。
      security: [{ bearerAuth: [] }]
      responses:
        "101": { description: Switching Protocols }
  /cli-auth/sessions:
    post:
      tags: [cli-auth]
      summary: 创建 CLI 设备授权会话（匿名，IP 限速 10/min）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceName, scopes]
              properties:
                deviceName:
                  { type: string, description: 设备名（hostname），≤64 字符 }
                scopes:
                  {
                    type: array,
                    items: { type: string },
                    description: 请求的 scope 列表,
                  }
      responses:
        "200":
          description: 会话创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId: { type: string }
                  secret:
                    {
                      type: string,
                      description: 轮询凭证（仅此一次返回，不进 URL）,
                    }
                  userCode: { type: string, description: 8 位短码 XXXX-XXXX }
                  authUrl: { type: string, description: 授权链接 }
                  pollInterval: { type: integer, default: 2 }
                  expiresAt: { type: integer }
                  scopes: { type: array, items: { type: string } }
        "400":
          $ref: "#/components/responses/Error"
        "429":
          description: 限速
  /cli-auth/sessions/{id}:
    get:
      tags: [cli-auth]
      summary: 查询授权会话信息（需登录 Cookie）
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 会话信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId: { type: string }
                  deviceName: { type: string }
                  requestedScopes: { type: array, items: { type: string } }
                  status:
                    {
                      type: string,
                      enum: [pending, approved, consumed, denied, expired],
                    }
                  createdAt: { type: integer }
                  expiresAt: { type: integer }
        "404":
          description: 会话不存在或已过期
  /cli-auth/sessions/by-code:
    get:
      tags: [cli-auth]
      summary: 按短码查询会话（需登录 Cookie，按用户限速 20/min）
      security: [{ cookieAuth: [] }]
      parameters:
        - name: code
          in: query
          required: true
          schema:
            { type: string, description: 8 位短码，大小写不敏感、容忍连字符 }
      responses:
        "200":
          description: 会话信息（同 /cli-auth/sessions/{id}）
        "404":
          description: 会话不存在或已过期
        "429":
          description: 限速
  /cli-auth/sessions/{id}/approve:
    post:
      tags: [cli-auth]
      summary: 批准授权（需登录 Cookie）
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scopes]
              properties:
                scopes:
                  {
                    type: array,
                    items: { type: string },
                    description: 授予的 scope（必须 ⊆ requestedScopes）,
                  }
      responses:
        "200":
          description: 批准成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: approved }
        "400":
          $ref: "#/components/responses/Error"
        "409":
          description: 终态冲突（会话不在 pending 状态）
  /cli-auth/sessions/{id}/deny:
    post:
      tags: [cli-auth]
      summary: 拒绝授权（需登录 Cookie）
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 拒绝成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: denied }
        "409":
          description: 终态冲突
  /cli-auth/sessions/{id}/poll:
    get:
      tags: [cli-auth]
      summary: 轮询会话状态（匿名，secret 门控）
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: secret
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 轮询结果（status 可为 pending/approved/denied/expired/consumed）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    {
                      type: string,
                      enum: [pending, approved, denied, expired, consumed],
                    }
                  apiKey:
                    {
                      type: string,
                      description: 仅 approved 时一次性返回明文 Key,
                    }
                  keyId: { type: string }
                  name: { type: string }
                  scopes: { type: array, items: { type: string } }
        "401":
          description: secret 无效
        "404":
          description: 会话不存在
        "429":
          description: 轮询过快
  /cli-auth/sessions/{id}/qr.txt:
    get:
      tags: [cli-auth]
      summary: 获取 ASCII 二维码（匿名，IP 限速 30/min）
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: ASCII 半块字符二维码（内容=authUrl）
          content:
            text/plain:
              schema: { type: string }
        "404":
          description: 会话不存在

  # --- Passkey 设备授权 ---
  /passkey-enroll/sessions:
    post:
      tags: [passkey-enroll]
      summary: 创建 Passkey 设备授权会话（需登录 Cookie）
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceName]
              properties:
                deviceName: { type: string, description: 发起设备名，≤64 字符 }
      responses:
        "200":
          description: 会话创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId: { type: string }
                  secret:
                    { type: string, description: 旧设备持有，不进二维码/链接 }
                  userCode: { type: string, description: 8 位短码 XXXX-XXXX }
                  authUrl: { type: string, description: 新设备接入链接 }
                  pollInterval: { type: integer, default: 2 }
                  expiresAt: { type: integer }
                  kind: { type: string, example: passkey }
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /passkey-enroll/sessions/{id}:
    get:
      tags: [passkey-enroll]
      summary: 匿名查询会话（IP 限速）
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 会话信息（最小视图，不含密钥）
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId: { type: string }
                  status:
                    {
                      type: string,
                      enum:
                        [
                          pending,
                          requested,
                          approved,
                          consumed,
                          denied,
                          expired,
                        ],
                    }
                  initiatorName: { type: string, description: 批准后才有值 }
                  deviceHint: { type: string, description: requested 后才有值 }
                  expiresAt: { type: integer }
        "404":
          description: 会话不存在或已过期
        "429":
          description: 限速
    delete:
      tags: [passkey-enroll]
      summary: 旧设备取消会话（需登录 Cookie）
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 取消成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: deleted }
        "401":
          $ref: "#/components/responses/Error"
        "404":
          description: 会话不存在
  /passkey-enroll/sessions/by-code:
    get:
      tags: [passkey-enroll]
      summary: 匿名按短码查询会话（IP 限速 10/min）
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 会话信息（同 /passkey-enroll/sessions/{id}）
        "404":
          description: 授权码无效或已过期
        "429":
          description: 限速
  /passkey-enroll/sessions/{id}/request:
    post:
      tags: [passkey-enroll]
      summary: 新设备敲门（匿名，IP 限速）
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceHint: { type: string, description: 新设备自报设备信息 }
      responses:
        "200":
          description: 敲门成功，返回 secret
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret: { type: string, description: 新设备轮询凭证 }
        "409":
          description: 非 pending 状态
        "429":
          description: 限速
  /passkey-enroll/sessions/{id}/watch:
    get:
      tags: [passkey-enroll]
      summary: 旧设备轮询会话状态（需登录 Cookie）
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 会话状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    {
                      type: string,
                      enum:
                        [
                          pending,
                          requested,
                          approved,
                          consumed,
                          denied,
                          expired,
                        ],
                    }
                  deviceHint: { type: string }
                  expiresAt: { type: integer }
        "401":
          $ref: "#/components/responses/Error"
        "404":
          description: 会话不存在
  /passkey-enroll/sessions/{id}/poll:
    get:
      tags: [passkey-enroll]
      summary: 新设备轮询（匿名，secret 门控）
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: secret
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 轮询结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    {
                      type: string,
                      enum:
                        [
                          pending,
                          requested,
                          approved,
                          consumed,
                          denied,
                          expired,
                        ],
                    }
                  attestationOptions:
                    { type: object, description: 仅 approved 时一次性返回 }
                  registrationToken:
                    { type: string, description: 仅 approved 时一次性返回 }
                  initiatorName: { type: string }
        "401":
          description: secret 无效
        "404":
          description: 会话不存在
        "429":
          description: 轮询过快
  /passkey-enroll/sessions/{id}/approve:
    post:
      tags: [passkey-enroll]
      summary: 旧设备批准（需登录 Cookie，requested→approved）
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 批准成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: approved }
        "401":
          $ref: "#/components/responses/Error"
        "409":
          description: 非 requested 状态
  /passkey-enroll/sessions/{id}/deny:
    post:
      tags: [passkey-enroll]
      summary: 旧设备拒绝（需登录 Cookie，requested→denied）
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 拒绝成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: denied }
        "401":
          $ref: "#/components/responses/Error"
        "409":
          description: 非 requested 状态
  /passkey-enroll/sessions/{id}/complete:
    post:
      tags: [passkey-enroll]
      summary: 新设备提交 attestation（匿名，registrationToken 门控）
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: registrationToken
          in: query
          required: true
          schema: { type: string }
        - name: name
          in: query
          required: false
          schema: { type: string, default: 新设备 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: 裸 WebAuthn attestation JSON
      responses:
        "200":
          description: 凭证创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  passkeyId: { type: string }
        "401":
          description: registrationToken 无效
        "409":
          description: 会话已完成或非 approved 状态
        "429":
          description: 限速
  /passkey-enroll/sessions/{id}/qr.txt:
    get:
      tags: [passkey-enroll]
      summary: 获取 ASCII 二维码（匿名，IP 限速）
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: ASCII 二维码（内容=authUrl）
          content:
            text/plain:
              schema: { type: string }
        "404":
          description: 会话不存在
  /keys/self:
    get:
      tags: [keys]
      summary: Key 自检（Bearer，任何有效 Key）
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Key 信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string }
                  prefix: { type: string }
                  scopes: { type: array, items: { type: string } }
        "401":
          $ref: "#/components/responses/Error"

  # ===== 管理后台（需登录会话 + 超管权限） =====
  /admin/stats:
    get:
      tags: [admin]
      summary: 系统总览统计（超管）
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: 系统级总览统计
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStats"
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: APIKey }
    cookieAuth: { type: apiKey, in: cookie, name: anotify_session }
  schemas:
    DeviceUpsert:
      type: object
      required: [endpoint, keys]
      properties:
        name: { type: string }
        platform: { type: string, enum: [ios, mac, win, android, other] }
        tags: { type: array, items: { type: string } }
        endpoint: { type: string }
        keys:
          type: object
          required: [p256dh, auth]
          properties:
            p256dh: { type: string }
            auth: { type: string }
        userAgent: { type: string }
    NotifyRequest:
      type: object
      required: [agentState, title]
      properties:
        agentId: { type: string }
        sessionId: { type: string }
        cwd: { type: string }
        agentState:
          { type: string, enum: [working, blocked, done, interrupted, error] }
        severity:
          { type: string, enum: [info, warning, error], description: 呈现语气；缺省由 agentState 派生 }
        durationMs: { type: integer }
        title: { type: string }
        body: { type: string }
        model: { type: string }
        link: { type: string }
        deviceTags:
          { type: array, items: { type: string }, description: 路由键；空=广播 }
        priority: { type: string, default: normal }
        ttl: { type: integer, default: 86400 }
    TestNotifyRequest:
      type: object
      description: 工作台测试通知请求体（会话鉴权，字段精简）
      properties:
        title: { type: string, default: 测试通知 }
        body: { type: string }
        agentState:
          type: string
          enum: [working, blocked, done, interrupted, error]
          default: done
        severity:
          { type: string, enum: [info, warning, error] }
        link: { type: string }
        deviceTags: { type: array, items: { type: string } }
        priority: { type: string, default: normal }
    SystemStats:
      type: object
      description: 管理后台系统总览统计
      properties:
        userCount: { type: integer }
        adminCount: { type: integer }
        activeUserCount:
          { type: integer, description: 近 7 天有消息上报的用户数 }
        messageCount: { type: integer }
        todayMessages: { type: integer }
        deviceCount: { type: integer }
        enabledDevices: { type: integer }
        keyCount: { type: integer }
        sessionCount: { type: integer, description: 活跃会话数 }
        deliveredCount: { type: integer }
        deliveryTotal: { type: integer }
        daily:
          type: array
          items:
            type: object
            properties:
              day: { type: string }
              count: { type: integer }
        topUsers:
          type: array
          items: { $ref: "#/components/schemas/UserMetric" }
    UserMetric:
      type: object
      properties:
        userId: { type: string }
        username: { type: string }
        displayName: { type: string }
        messageCount: { type: integer }
        deviceCount: { type: integer }
...
  responses:
    Error:
      description: 错误
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }